requirements.txt
celery
aiohttp
dash
dash-bootstrap-components
db-sqlite3
polygon
mysql